Move hasNoCertIssuers
